Move form related workout files into route group for organization and shared loading.tsx
